Cleaned up the root (src) directory
