+++ src/client/src/ui/common/popout/popoutEpic.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 import { createAction } from 'redux-actions'
-import { ACTION_TYPES as REGIONS_ACTIONS } from './regions/regionsOperations'
-import { ACTION_TYPES as TILE_ACTIONS, tileUndocked } from './ui/spotTile/actions'
-import { getPopoutService } from './ui/common/popout'
+import { ACTION_TYPES as REGIONS_ACTIONS } from '../../../regions/regionsOperations'
+import { ACTION_TYPES as TILE_ACTIONS, tileUndocked } from '../../spotTile/actions'
+import { getPopoutService } from './index'
 import { Provider } from 'react-redux'
 import { combineEpics } from 'redux-observable'
